% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_exerc.R
\name{gen_exerc}
\alias{gen_exerc}
\title{Gerar exercícios}
\usage{
gen_exerc(
  exs,
  curso,
  semestre,
  n_exerc,
  template = c("exercicio_pt.tex", "solution_pt.tex"),
  decimal = ","
)
}
\arguments{
\item{exs}{vetor com o nome dos arquivos dos exercícios}

\item{curso}{Nome da disciplina}

\item{semestre}{Semestre da disciplina}

\item{n_exerc}{numero de exercicios que serao gerados}

\item{template}{arquivo latex com o template}

\item{decimal}{separador de decimal}
}
\description{
Gerar exercícios
}
