% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masc.R
\name{masc}
\alias{masc}
\title{masc Multiple Answer Single Choice Question}
\usage{
masc(statements, statements_truth, explanation = NULL, n_wrong = 4)
}
\arguments{
\item{statements}{A list containing the statements.Each element of the list
can be a vector, in which only one of the elements of the vector will be
randomly selected to compose the question.}

\item{statements_truth}{a list of logical values (TRUE/FALSE) of statements}

\item{explanation}{A list with the explanation (optional).}

\item{n_wrong}{Number of wrong alternatives. 4 by default.}
}
\value{
A list with the randomized statements (character),
the alternatives (character) and the answers (logical).
}
\description{
Presents several statements. Judge each of them as correct or incorrect.
Tick the true alternative.
}
\examples{
statements <- list(c("A"), c("D", "E"), c("G", "H", "I"))
statements_truth <- list(c(TRUE), c(FALSE, TRUE), c(FALSE, FALSE, TRUE))
explanation <- list(c("A is a vowel"), c("D is a consonant", "E is a vowel"),
 c("G is a consonant", "H is a consonant", "I is a vowel"))
masc(statements, statements_truth, explanation)
}
